Created view transaction function
This function allows the user to filter and display previous transactions by date range and optionally by category. User inputs are converted to datetime so that they can be compared with the transaction dates in the .json and make sure they fall within the user's indicated range. I also built some "fail-safe" features to ensure that incorrect input or null returns don't break the code (making sure the input date format is correct). I've also built in some formatting to make sure the output is meaningful or indicates if no transactions are found for their given inputs.